Fixed a few more warnings.

+++ 2048 FB/Board+ModelLayer02.m
@@ -106,7 +106,7 @@
 }
 
 -(BOOL)setIntegerScore: (NSInteger) score {
-	self.score = [NSDecimalNumber decimalNumberWithString:[NSString stringWithFormat:@"%ld", score]];
+	self.score = [NSDecimalNumber decimalNumberWithString:[NSString stringWithFormat:@"%ld", (long)score]];
 	return YES;
 }
